@@ -10,29 +10,27 @@
         "call",
         "callable"
     ],
-    "homepage": "https://github.com/IcecaveStudios/evoke",
+    "homepage": "https://github.com/icecave/evoke",
     "license": "MIT",
     "authors": [
         {
             "name": "James Harris",
-            "email": "james.harris@icecave.com.au",
+            "email": "mailjamesharris@gmail.com",
             "homepage": "https://github.com/jmalloc"
         }
     ],
     "require": {
-        "php": ">=5.3"
+        "php": "^7.1"
     },
     "require-dev": {
-        "icecave/archer": "~1"
+        "phpunit/phpunit": "^7"
     },
-    "autoload": {
-        "psr-4": {
-            "Icecave\\Evoke\\": "src"
-        }
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-develop": "1.0.x-dev"
+    "autoload": { "psr-4": { "Icecave\\Evoke\\": "src" } },
+    "autoload-dev": { "psr-4": { "Icecave\\Evoke\\": "test/src" } },
+    "config": {
+        "sort-packages": true,
+        "platform": {
+            "php": "7.1"
         }
     }
 }
